refactor(pipeline): extract ImagingField base class
Split nircam/field.py (~625 lines) into a generic ImagingField base in
common/imaging/field.py and a thin NircamField subclass with only the
NIRCam-specific bits (bad-pixel / wisp reference subdirectories, SW/LW
filter classifier).

The base class owns the TOML loader, workspace setup, file-glob
discovery (uncals + canonical exposures), and tile WCS geometry. It uses
ClassVar attributes INSTRUMENT_NAME, DETECTOR_SHAPE, KNOWN_STEPS that
subclasses set to drive instrument-specific path components and step
override parsing. A _setup_instrument_workspace() hook lets subclasses
add their own reference subdirectories (NIRCam: bad_pixels + wisps;
MIRI will add warm_pixels and persistence caches).

The ``Field`` name is kept as a module-level alias for NircamField so
existing import sites (``from campfire_pipeline.nircam.field import Field``)
continue to work without churn.

The has_step fallback in get_exposure_files (when no StepStatus cache is
passed) now uses common.cfp.has_step, which is instrument-agnostic after
the Step 2 refactor — no per-instrument cfp import needed in the base
class.

All 91 existing tests pass, including the 15 RGB tests that exercise
Field.load() with various fields.toml layouts. CLI smoke (cfpipe nircam
--help) passes.
